change counter variable range

+++ prime/main-xxx.c
@@ -2,7 +2,7 @@
 
 int main(int argc, char *argv[])
 {
-    long long int i = 0;
+    unsigned long int i = 0;
 
     for (i = 0; i < 1055125; i++)
     {
